delete unused files and remove build warnings

+++ server/reactor/echo_client_handler.h
@@ -32,7 +32,7 @@ Accept: text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/a
 Accept-Encoding: gzip, deflate\r\n\
 Accept-Language: zh-CN,zh;q=0.9,en-US;q=0.8,en;q=0.7";
        int ret = this->write(data, ::strlen(data));
-       if(ret != strlen(data))
+       if(ret != static_cast<int64_t>(strlen(data)))
        {
            LOG(WARNING) << "write error";
            return -1;
